edit commit files mistake
